refactor: normalize website structure
