+++ packages/website/src/pages/docs/integration/cdn-setup/page.tsx
